Core: Update prototypes to suite headers prototypes

+++ libft/srcs/core/ft_memchr.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void	*ft_memchr(const void *s, int c, unsigned long n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned char *s2;
