+++ libft/srcs/core/ft_memcmp.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_memcmp(const void *s1, const void *s2, unsigned long n)
+int		ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	unsigned char str1;
 	unsigned char str2;
